Partial crash fix for high resolutions
This occured much more often in TPI but I manaed to reproduce it on TPW
as well. The main issue is a small buffer overflow (4 bytes) occuring in
functions `0058F190` for TPW and `007043A0` for TPI. When the 16KB
buffer is full, the `0` integer value is written just after the buffer,
overwriting the return address of the calling function. So, instead of
returning to the previous function, the EIP register is set to `0` and
the game crashes.

Increasing the buffer of just 4 bytes could have solved this crash but
this wouldn't have solved the root cause of the issue which is that the
buffer was full. So I instead increased the buffer size by a large
amount and applied the same fix to other functions.

With this fix, instead of crashing 100% of the times on certain frames,
it now crashes 95% of the times and 1 frame later. When it does not
crash, some textures may be heavily corrupted.

There are still other issues remaining before claiming these issues
completely fixed.